Rename ProseList to ListGenerator
Add component documentation to explain behavior and show an example.
